+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -129,6 +129,13 @@ New checks
 
   Finds constructors which just initialize class members, so they can be
   replaced with in-class member initializers.
+- New :doc:`misc-avoid-std-io-outside-main
+  <clang-tidy/checks/misc-avoid-std-io-outside-main>` check.
+
+  Finds predefined standard stream objects like ``cin``, ``wcin``, ``cout``,
+  ``wcout``, ``cerr`` or ``wcerr``, that are used outside the ``main``
+  function. It also finds uses of ``cstdio`` and ``stdio.h`` functions like
+  ``printf()`` outside the ``main`` function.
 
 New check aliases
 ^^^^^^^^^^^^^^^^^